changed sample_for_prop method to be contained in wind field generator class.
*moved sample_for_prop
